🧐 Refactor FoodTruck domain class structure and add smart enums
Deleted the Class1 file in the FoodTruck.Domain as it was not needed. Created separate files for each of the new classes - MobileFoodFacility, FacilityType, and Status. These include a new detailed MobileFoodFacility class and two smart enums for FacilityType and Status. Additionally, the Ardalis.SmartEnum package has been added to the domain project for advanced enum capabilities.