🚧 Add infrastructure layer and update dependencies
This commit introduces the infrastructure layer with the addition of the MobileFoodFacilitiesRepository and FoodTruckDbContext classes. The commit also updates the dependency injection to include newly created components. Changes to the FoodTruck.Api project include the addition of the infrastructure services and declaration of the project reference in the csproj file.